requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
peft>=0.9.0
safetensors>=0.4.2
bitsandbytes>=0.42.0
insightface>=0.7.3
onnxruntime-gpu>=1.17.0
opencv-python>=4.9.0
Pillow>=10.0.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
tqdm>=4.66.0
wandb>=0.16.0
open-clip-torch>=2.24.0
pytorch-fid>=0.3.0
gdown>=5.1.0
fawkes>=1.0.4
pytest>=8.0.0
